+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/checkers/BindingContainerCallableChecker.kt
@@ -144,6 +144,15 @@ internal object BindingContainerCallableChecker :
       return
     }
 
+    declaration
+      .getAnnotationByClassId(DaggerSymbols.ClassIds.DAGGER_REUSABLE_CLASS_ID, session)
+      ?.let {
+        reporter.reportOn(it.source ?: source, MetroDiagnostics.DAGGER_REUSABLE_ERROR)
+        return
+      }
+
+    // After the reusable check because reusable is technically a scope and we don't want to
+    // double-report
     if (annotations.isBinds && annotations.scope != null) {
       reporter.reportOn(
         annotations.scope.fir.source ?: source,
@@ -151,13 +160,6 @@ internal object BindingContainerCallableChecker :
         "@Binds declarations may not have scopes.",
       )
     }
-
-    declaration
-      .getAnnotationByClassId(DaggerSymbols.ClassIds.DAGGER_REUSABLE_CLASS_ID, session)
-      ?.let {
-        reporter.reportOn(it.source ?: source, MetroDiagnostics.DAGGER_REUSABLE_ERROR)
-        return
-      }
 
     if (declaration.typeParameters.isNotEmpty()) {
       val type = if (annotations.isProvides) "Provides" else "Binds"